Combine class for pre and post mission groups. Other cleanup.
